requirements.txt
tensorflow
tensorflow_addons
pillow
pygame